MMVv2: Move InstanceStoreV1 into InstanceV1 class
